backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
playwright==1.47.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
httpx==0.27.2
